Math Tutor 2
Made to ask the user for one of 5 categories (+ , - , * , / , Quit). Which then randomly generates 5 problems and tallies up the correct answers for when the user quits, calculating and displaying the final results in percentage format ( I.E. if 2 out of 5 correct, it displays your score as 40). Each correct answer giving a compliment and each incorrect answer giving encouragement. Both of which are generated based on random selection.